Tons of refactoring to make gene function parsing more modular and reusable
