+++ src/tests/java/org/broadinstitute/dropseqrna/utils/TestUtils.java
@@ -24,6 +24,7 @@
 
 package org.broadinstitute.dropseqrna.utils;
 
+import htsjdk.samtools.SAMFileHeader;
 import htsjdk.samtools.SAMRecordIterator;
 import htsjdk.samtools.SamReader;
 import htsjdk.samtools.SamReaderFactory;
@@ -93,7 +94,9 @@ public class TestUtils {
 		final SamReader expectedReader = SamReaderFactory.makeDefault().open(expected);
 		final SamReader actualReader = SamReaderFactory.makeDefault().open(actual);
 		try {
-			Assert.assertEquals(expectedReader.getFileHeader(), actualReader.getFileHeader());
+			SAMFileHeader eh= expectedReader.getFileHeader();
+			SAMFileHeader ah= actualReader.getFileHeader();			
+			Assert.assertEquals(eh, ah);
 			assertSamRecordsSame(actual, expected, expectedReader, actualReader);
 		} finally {
 			CloserUtil.close(expectedReader);
